Parameters now indexed and updated properly.
It's kind of ugly though; maybe better to move parameter updates to the input layer --- only that would separate neurons from their parameters, which might be a bit opaque. It feels like a question of ownership; which part of the network _owns_ the parameters: individual neurons or the network as a whole?